+++ src/Runtime/__mem.h
@@ -3,8 +3,8 @@
 
 #include <platform.h>
 
-SECTION_INIT void* memset(void* dst, int val, unsigned long n);
-SECTION_INIT void __fill_mem(void* dst, int val, unsigned long n);
-SECTION_INIT void* memcpy(void* dst, const void* src, unsigned long n);
+SECTION_INIT void* memset(void* dst, int val, size_t n);
+SECTION_INIT void __fill_mem(void* dst, int val, size_t n);
+SECTION_INIT void* memcpy(void* dst, const void* src, size_t n);
 
 #endif
